allow other actions when tapping on blank space
With this, the plugin will only handle tapping on text.
Taps on areas without text will be handled as usual. This allows page
flipping, toggling the status bar, tapping on corners etc.
Introducing slight margins to ensure tapping the top corners is possible
(at least at the very edge), however the bottom corners might be in
conflict with the status bar or the bottom menu.